packages/api/pyproject.toml
[project]
name = "navaro-api"
version = "1.0.0"
description = "Navaro API - Sistema de agendamento para barbearias e salões"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.17",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    
    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # HTTP Client
    "httpx>=0.28.0",
    
    # Cache & Queue
    "redis>=5.2.0",
    
    # Background Tasks
    "celery[redis]>=5.4.0",
    
    # External Services
    "stripe>=11.0.0",
    "twilio>=9.4.0",
    "boto3>=1.35.0",
    
    # Utilities
    "phonenumbers>=8.13.0",
    "python-slugify>=8.0.0",
    "pillow>=11.0.0",
    "qrcode[pil]>=7.4.0",
    
    # Logging
    "structlog>=24.4.0",
    
    # Retry
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
    
    # Linting & Formatting
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    
    # Type stubs
    "types-passlib>=1.7.0",
    "types-python-jose>=3.3.0",
]

[project.scripts]
navaro = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "E712",   # comparison to True/False
    "RUF001", # ambiguous character
    "RUF015", # prefer next over slice
    "PT003",  # pytest fixture scope implied
    "ARG",    # unused arguments (common in interfaces/tests)
    "UP042",  # prefer StrEnum (keep existing style for now)
    "N805",   # cls as first arg in pydantic validators
    "F403",   # star imports
    "N818",   # exception naming
    "W293",   # whitespace in multiline strings
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S", "PLR2004", "SIM", "F841"]
"__init__.py" = ["F401"]


[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "-ra",
    "-q",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
